Move transaction to separate module
